agents/vh-faq-agent/livekit-agent/requirements.txt
livekit-agents~=1.1.1
livekit-plugins-openai~=1.1.1
livekit-plugins-deepgram~=1.1.1
livekit-plugins-silero~=1.1.1
livekit-plugins-turn-detector~=1.1.1
python-dotenv~=1.1.0
aiohttp~=3.12.13
livekit-plugins-noise-cancellation~=0.2.4